+++ src/nationalarchives/components/header/header.stories.js
@@ -125,15 +125,17 @@ Desktop.args = {
   classes: "tna-header--demo",
 };
 Desktop.play = async ({ canvasElement }) => {
-  const navigationItems = canvasElement.querySelector(
+  await new Promise((r) => setTimeout(r, 100));
+
+  const $navigationItems = canvasElement.querySelector(
     `.tna-header__navigation-items`,
   );
-  const navigationToggle = canvasElement.querySelector(
+  const $navigationToggle = canvasElement.querySelector(
     `.tna-header__navigation-toggle-button`,
   );
 
-  await expect(navigationItems).toBeVisible();
-  await expect(navigationToggle).not.toBeVisible();
+  await expect($navigationItems).toBeVisible();
+  await expect($navigationToggle).not.toBeVisible();
 };
 
 export const Mobile = Template.bind({});
@@ -164,46 +166,89 @@ Mobile.args = {
   classes: "tna-header--demo",
 };
 Mobile.play = async ({ args, canvasElement, step }) => {
+  await new Promise((r) => setTimeout(r, 100));
+
   const canvas = within(canvasElement);
 
-  const navigationItems = canvasElement.querySelector(
+  const $navigationItems = canvasElement.querySelector(
     `.tna-header__navigation-items`,
   );
-  const [linkA, linkB, linkC] = args.navigation.map((navigationItem) =>
+  const [$linkA, $linkB, $linkC] = args.navigation.map((navigationItem) =>
     canvas.getByText(navigationItem.text),
   );
-  const navigationToggle = canvasElement.querySelector(
+  const $navigationToggle = canvasElement.querySelector(
     `.tna-header__navigation-toggle-button`,
   );
 
   await step("Initial load", async () => {
-    await expect(navigationItems).not.toBeVisible();
-    await expect(navigationToggle).toBeVisible();
-    await expect(navigationToggle).toHaveAttribute("aria-expanded", "false");
-    await expect(linkA).not.toBeVisible();
-    await expect(linkB).not.toBeVisible();
-    await expect(linkC).not.toBeVisible();
+    await expect($navigationItems).not.toBeVisible();
+    await expect($navigationToggle).toBeVisible();
+    await expect($navigationToggle).toHaveAttribute("aria-expanded", "false");
+    await expect($linkA).not.toBeVisible();
+    await expect($linkB).not.toBeVisible();
+    await expect($linkC).not.toBeVisible();
   });
 
   await step("Open the menu", async () => {
-    await userEvent.click(navigationToggle);
-    await expect(navigationItems).toBeVisible();
-    await expect(navigationToggle).toBeVisible();
-    await expect(navigationToggle).toHaveAttribute("aria-expanded", "true");
-    await expect(linkA).toBeVisible();
-    await expect(linkB).toBeVisible();
-    await expect(linkC).toBeVisible();
+    await userEvent.click($navigationToggle);
+    await expect($navigationItems).toBeVisible();
+    await expect($navigationToggle).toBeVisible();
+    await expect($navigationToggle).toHaveAttribute("aria-expanded", "true");
+    await expect($linkA).toBeVisible();
+    await expect($linkB).toBeVisible();
+    await expect($linkC).toBeVisible();
   });
 
   await step("Close the menu", async () => {
-    await userEvent.click(navigationToggle);
-    await expect(navigationItems).not.toBeVisible();
-    await expect(navigationToggle).toBeVisible();
-    await expect(navigationToggle).toHaveAttribute("aria-expanded", "false");
-    await expect(linkA).not.toBeVisible();
-    await expect(linkB).not.toBeVisible();
-    await expect(linkC).not.toBeVisible();
+    await userEvent.click($navigationToggle);
+    await expect($navigationItems).not.toBeVisible();
+    await expect($navigationToggle).toBeVisible();
+    await expect($navigationToggle).toHaveAttribute("aria-expanded", "false");
+    await expect($linkA).not.toBeVisible();
+    await expect($linkB).not.toBeVisible();
+    await expect($linkC).not.toBeVisible();
   });
 
-  navigationToggle.blur();
+  $navigationToggle.blur();
 };
+
+// export const NoJavascript = Template.bind({});
+// NoJavascript.args = {
+//   navigation: [
+//     {
+//       text: "Alpha",
+//       href: "#/alpha",
+//       selected: true,
+//     },
+//     {
+//       text: "Beta",
+//       href: "#/beta",
+//     },
+//     {
+//       text: "Gamma",
+//       href: "#/gamma",
+//     },
+//   ],
+//   classes: "tna-header--demo",
+// };
+// NoJavascript.play = async ({ args, canvasElement, step }) => {
+//   document.documentElement.classList.remove("tna-template--js-enabled");
+//   canvasElement.innerHTML = canvasElement.innerHTML;
+//   const $navigationToggleWrapper = canvasElement.querySelector(
+//     `.tna-header__navigation-toggle`,
+//   );
+//   $navigationToggleWrapper.innerHTML = "";
+
+//   const canvas = within(canvasElement);
+
+//   // const $navigationItems = canvasElement.querySelector(
+//   //   `.tna-header__navigation-items`,
+//   // );
+//   // const [$linkA, $linkB, $linkC] = args.navigation.map((navigationItem) =>
+//   //   canvas.getByText(navigationItem.text),
+//   // );
+//   // await expect($navigationItems).toBeVisible();
+//   // await expect($linkA).toBeVisible();
+//   // await expect($linkB).toBeVisible();
+//   // await expect($linkC).toBeVisible();
+// };